+++ generators/python/src/fern_python/generators/sdk/declaration_referencers/__init__.py
@@ -2,6 +2,9 @@ from .environments_enum_declaration_referencer import (
     EnvironmentsEnumDeclarationReferencer,
 )
 from .error_declaration_referencer import ErrorDeclarationReferencer
+from .oauth_token_provider_declaration_referencer import (
+    OAuthTokenProviderDeclarationReferencer,
+)
 from .root_client_declaration_referencer import RootClientDeclarationReferencer
 from .subpackage_async_client_declaration_referencer import (
     SubpackageAsyncClientDeclarationReferencer,
@@ -18,4 +21,5 @@ __all__ = [
     "TypeDeclarationReferencer",
     "RootClientDeclarationReferencer",
     "SubpackageAsyncClientDeclarationReferencer",
+    "OAuthTokenProviderDeclarationReferencer",
 ]