+++ seed/python-sdk/auth-environment-variables/src/seed/client.py
@@ -43,7 +43,7 @@ class SeedAuthEnvironmentVariables:
         x_another_header: typing.Optional[str] = os.getenv("ANOTHER_ENV_VAR"),
         api_key: typing.Optional[str] = os.getenv("FERN_API_KEY"),
         timeout: typing.Optional[float] = None,
-        follow_redirects: typing.Optional[bool] = None,
+        follow_redirects: typing.Optional[bool] = True,
         httpx_client: typing.Optional[httpx.Client] = None
     ):
         _defaulted_timeout = timeout if timeout is not None else 60 if httpx_client is None else None
@@ -100,7 +100,7 @@ class AsyncSeedAuthEnvironmentVariables:
         x_another_header: typing.Optional[str] = os.getenv("ANOTHER_ENV_VAR"),
         api_key: typing.Optional[str] = os.getenv("FERN_API_KEY"),
         timeout: typing.Optional[float] = None,
-        follow_redirects: typing.Optional[bool] = None,
+        follow_redirects: typing.Optional[bool] = True,
         httpx_client: typing.Optional[httpx.AsyncClient] = None
     ):
         _defaulted_timeout = timeout if timeout is not None else 60 if httpx_client is None else None