+++ seed/python-sdk/oauth-client-credentials/src/seed/core/client_wrapper.py
@@ -8,7 +8,14 @@ from .http_client import AsyncHttpClient, HttpClient
 
 
 class BaseClientWrapper:
-    def __init__(self, *, base_url: str, timeout: typing.Optional[float] = None):
+    def __init__(
+        self,
+        *,
+        token: typing.Optional[typing.Union[str, typing.Callable[[], str]]] = None,
+        base_url: str,
+        timeout: typing.Optional[float] = None,
+    ):
+        self._token = token
         self._base_url = base_url
         self._timeout = timeout
 
@@ -18,7 +25,16 @@ class BaseClientWrapper:
             "X-Fern-SDK-Name": "fern_oauth-client-credentials",
             "X-Fern-SDK-Version": "0.0.1",
         }
+        token = self._get_token()
+        if token is not None:
+            headers["Authorization"] = f"Bearer {token}"
         return headers
+
+    def _get_token(self) -> typing.Optional[str]:
+        if isinstance(self._token, str) or self._token is None:
+            return self._token
+        else:
+            return self._token()
 
     def get_base_url(self) -> str:
         return self._base_url
@@ -28,12 +44,26 @@ class BaseClientWrapper:
 
 
 class SyncClientWrapper(BaseClientWrapper):
-    def __init__(self, *, base_url: str, timeout: typing.Optional[float] = None, httpx_client: httpx.Client):
-        super().__init__(base_url=base_url, timeout=timeout)
+    def __init__(
+        self,
+        *,
+        token: typing.Optional[typing.Union[str, typing.Callable[[], str]]] = None,
+        base_url: str,
+        timeout: typing.Optional[float] = None,
+        httpx_client: httpx.Client,
+    ):
+        super().__init__(token=token, base_url=base_url, timeout=timeout)
         self.httpx_client = HttpClient(httpx_client=httpx_client)
 
 
 class AsyncClientWrapper(BaseClientWrapper):
-    def __init__(self, *, base_url: str, timeout: typing.Optional[float] = None, httpx_client: httpx.AsyncClient):
-        super().__init__(base_url=base_url, timeout=timeout)
+    def __init__(
+        self,
+        *,
+        token: typing.Optional[typing.Union[str, typing.Callable[[], str]]] = None,
+        base_url: str,
+        timeout: typing.Optional[float] = None,
+        httpx_client: httpx.AsyncClient,
+    ):
+        super().__init__(token=token, base_url=base_url, timeout=timeout)
         self.httpx_client = AsyncHttpClient(httpx_client=httpx_client)