+++ seed/ruby-sdk/oauth-client-credentials-environment-variables/test/test_fern_oauth_client_credentials_environment_variables.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require_relative "test_helper"
+require "fern_oauth_client_credentials_environment_variables"
+
+# Basic SeedOauthClientCredentialsEnvironmentVariablesClient tests
+class TestSeedOauthClientCredentialsEnvironmentVariablesClient < Minitest::Test
+  def test_function
+    # SeedOauthClientCredentialsEnvironmentVariablesClient::Client.new
+  end
+end