+++ seed/ts-express/oauth-client-credentials-environment-variables/errors/index.d.ts
@@ -1,0 +1,1 @@
+export { SeedOauthClientCredentialsEnvironmentVariablesError } from "./SeedOauthClientCredentialsEnvironmentVariablesError";