+++ seed/ts-express/oauth-client-credentials-environment-variables/errors/index.js
@@ -1,0 +1,5 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.SeedOauthClientCredentialsEnvironmentVariablesError = void 0;
+var SeedOauthClientCredentialsEnvironmentVariablesError_1 = require("./SeedOauthClientCredentialsEnvironmentVariablesError");
+Object.defineProperty(exports, "SeedOauthClientCredentialsEnvironmentVariablesError", { enumerable: true, get: function () { return SeedOauthClientCredentialsEnvironmentVariablesError_1.SeedOauthClientCredentialsEnvironmentVariablesError; } });