+++ seed/ts-sdk/oauth-client-credentials-environment-variables/src/index.ts
@@ -1,0 +1,6 @@
+export * as SeedOauthClientCredentialsEnvironmentVariables from "./api";
+export { SeedOauthClientCredentialsEnvironmentVariablesClient } from "./Client";
+export {
+    SeedOauthClientCredentialsEnvironmentVariablesError,
+    SeedOauthClientCredentialsEnvironmentVariablesTimeoutError,
+} from "./errors";