+++ generators/python/fastapi/CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.9.0-rc1] - 2024-04-22
+
+- Feature: The generator now depends on v39 of Intermediate Representation which requires the latest
+  CLI.
+
 ## [0.9.0-rc0]
 
-- Feature: The generator now depends on v38 of Intermediate Representation which requires the latest 
+- Feature: The generator now depends on v38 of Intermediate Representation which requires the latest
   CLI.
 
 - Improvement: Remove support for Python 3.7. In order to support newer versions of libraries we depend on (such as typing and typing-extensions), we must move on to Python 3.8. With this change we are also able to: